wip(build): replace build system
After much struggles with the previous build system, I've decided to
start over. Matt left it in a suitable state for a single-stage build
for a BSD/Linux-native DCC for making 68k ROMable objects, but I've been
unable to adapt it for a full multi-stage, Amiga-targetted build.

d(x)make is unfortunately not up to the task, so I've started again
using GNU Make. I did consider BSD Make as a hat-tip to Matt's recent
work in the world of BSDs, but in the end I went for the system that I
know the best and which is usually installed by default.