fix(dc1): #ifdef out quad word implementation
I'm not sure where this has come from, it seems to have been added
relatively recently in DICE history, and is odd because DICE explicitly
does *not* support anything bigger than 32-bit words. It may be related
to "C-to-D" support - something else which is odd and may have been
added recently to DICE.

Marking the code with #ifdef to omit it from the compilation, and
therefore allow compilation on Amigas to work again.