fix(dcc): use registered libraries with -mr option given
This was an interesting one. DICE is documented to have -mr, -mR and
-mRR options. But at some point in history, dc1 was modified so that all
three of these options behave the same (the old behaviour at least still
partly behind #ifdefs). However dcc still treats -mr different to -mR
and -mRR: when -mr is given, it would still link with the stack-args
version of c.lib and amiga.lib. This would effectively break all builds
(unless they never called c.lib/amiga.lib). This did not seem to make
sense to me and I can't think of a valid use case, so I've changed this
behaviour of -mr to use the registered-args libraries like -mR and -mRR.