fix(dc1): fix for LEXIDX errors in das
Occasionally, das would fail with errors about a LEXIDX keyword in its
input file. I'm not quite sure what the trigger for those errors is. The
cause is a `printf` in dc1 which emits some runtime debugging
information; however, in dc1, stdout is redirected to the output file
that goes to das, so this debug message was inadvertantly ending up
mixed in to the produced assembly. Removed this debugging message and
the LEXIDX problem seems to be gone.